Predictive ML Model for mode of transport
This project requires one to understand what mode of transport employees prefers to commute to their office. The attached dataset Cars_edited.csv includes employee information about their mode of transport as well as their personal and professional details like age, salary, work exp. We need to predict whether or not an employee will use Car as a mode of transport. Also, which variables are a significant predictor behind this decision?